requirements.txt
Django==2.2.0
djangorestframework
django-filter
fasttext
